feat: verify user otp flow

+++ data/userVerification.js
@@ -1,15 +1,33 @@
 const { ObjectId } = require("mongodb");
-const { helpers } = require("../utils/helpers");
+const { validations, helpers, checkId } = require("../utils/helpers");
 const { UserOTPVerification } = require("./models/userOTPVerification.model");
 const {
   userVerificationCollection,
   usersCollection,
 } = require("../config/mongoCollections");
+const bcrypt = require("bcryptjs");
 
 const createUserVerification = async (verifyObj) => {
-  // TODO validations
+  verifyObj.userId = checkId(verifyObj.userId, "User ID");
+
+  const usersDB = await usersCollection();
+  const userById = await usersDB.findOne({ _id: ObjectId(verifyObj.userId) });
+  if (userById === null) throw new Error("No user with that id");
+
+  if (!helpers.isNumberValid(verifyObj?.otp)) {
+    throw new Error("OTP must be a number");
+  }
+  if (!validations.isOTPValid(verifyObj?.otp)) {
+    throw new Error("Invalid OTP generated. Try again.");
+  }
+  const hashedOtp = await bcrypt.hash(
+    helpers.sanitizeString("" + verifyObj?.otp),
+    10
+  );
+  const newObj = new UserOTPVerification(verifyObj.userId, hashedOtp);
+
   const userVerificationDB = await userVerificationCollection();
-  const insertInfo = await userVerificationDB.insertOne(verifyObj);
+  const insertInfo = await userVerificationDB.insertOne(newObj);
   if (!insertInfo.acknowledged || !insertInfo.insertedId)
     throw "Could not create user verification entry";
 
@@ -18,16 +36,28 @@ const createUserVerification = async (verifyObj) => {
 };
 
 const getUserVerificationByUserId = async (userId) => {
-  // TODO validations
+  userId = checkId(userId, "User ID");
+
+  const usersDB = await usersCollection();
+  const userById = await usersDB.findOne({ _id: ObjectId(userId) });
+  if (userById === null) throw new Error("No user with that id");
+  // const userById = await userDL.getUserById(userId); // this will throw error if no user found
+
   const userVerificationDB = await userVerificationCollection();
-  const verifications = await userVerificationDB.findOne({
+  const verification = await userVerificationDB.findOne({
     userId: ObjectId(userId),
   });
-  return verifications?.map((x) => new UserOTPVerification().deserialize(x));
+  return new UserOTPVerification().deserialize(verification);
 };
 
 const deleteMany = async (userId) => {
-  // TODO validations
+  userId = checkId(userId, "User ID");
+
+  const usersDB = await usersCollection();
+  const userById = await usersDB.findOne({ _id: ObjectId(userId) });
+  if (userById === null) throw new Error("No user with that id");
+  // const userById = await userDL.getUserById(userId); // this will throw error if no user found
+
   const userVerificationDB = await userVerificationCollection();
   const verifications = await userVerificationDB.deleteMany({
     userId: ObjectId(userId),
